Move the dynamic columns functions to separate file
